Tidy up buildings module
